requirements.txt
Django==5.0.3
djangorestframework==3.15.1
bcrypt==4.1.2